Fixed Some Merging Errors
CRS file was not compatible with the length, perimeter, and area functions for Polsby-Popper. So, I added a conversion to UTM Zone 17N(Most of florida, most accurate available) to do the calculations.

Also added a quick change from unary_union to geometry.union_all() because unary_union will be deprecated soon.